+++ Front-End/projectfe/src/components/NavBar.js
@@ -16,7 +16,7 @@ export const NavBar =() => {
             <Nav.Link href="/home">Home</Nav.Link>
             <Nav.Link href="/about">About</Nav.Link>
             <Nav.Link href="/doctorlog">Doctor</Nav.Link>
-            <Nav.Link href="/patientloging">Patient</Nav.Link>
+            <Nav.Link href="/patientlog">Patient</Nav.Link>
             <Nav.Link href="/contact">Contact</Nav.Link>
           </Nav>
         </Navbar.Collapse>
