+++ Front-End/projectfe/src/pages/doctorReg.css
@@ -1,15 +1,15 @@
-/*.patient{
+.regdoctor{
     display: flex;
     align-items: center;
     justify-content: center;
     height: 60vh;
     background: #D3EAF6;
 }
-form{
+.regform{
     background:#D3EAF6;
     padding: 0px 60px;
 }
-.BTN{
+.BTN1{
     background: #2760F1;
     color: white;
     font-weight: bold;
@@ -21,14 +21,14 @@ form{
     font-size: 20px;
     top: 420px;
 }
-.P1{
+.D1{
     color: #2760F1;
     text-align: center;
     position: absolute;
     top: 100px;
     left: 590px;
 }
-.P2{
+.D2{
     color: black;
     text-align: center;
     position: absolute;
@@ -36,7 +36,7 @@ form{
     left: 20px;
     font-weight: bold;
 }
-.P3 {
+.D3 {
     color: black;
     text-align: center;
     position: absolute;
@@ -44,7 +44,7 @@ form{
     left: 600px;
     font-weight: bold;
 }
-.Link{
+.regLink1{
     color: blue;
     font-size: 18px;
     text-align: center;
@@ -61,6 +61,4 @@ form{
         visibility: collapse;
     }
 
-
-
-}*/
+}
