+++ Front-End/projectfe/src/pages/doctorlog.js
@@ -3,6 +3,7 @@ import React, { useState } from "react";
 import loginImg from "../components/loging/doctor.jpg";
 import "../components/loging/styles.css";
 import "bootstrap/dist/css/bootstrap.min.css";
+import {Link} from 'react-router-dom';
 
 export const Doctor = () => {
   return (
@@ -53,7 +54,7 @@ export const Doctor = () => {
             <p>Forgot Password</p>
             </div>
             <div className="register-here">
-            <p>Don't have an account? <a className="register-hereLink" href="signup.html">REGISTER HERE</a></p>
+            <p>Don't have an account? <Link to ="/doctor-reg" className="register-hereLink">REGISTER HERE</Link></p>
             </div>
             </div>
           </form>
