+++ Front-End/projectfe/src/pages/patientlog.js
@@ -1,90 +1,65 @@
+import React, { useState } from "react";
+import loginImg from "../components/loging/patient.jpg";
+import "../components/loging/styles.css";
+import "bootstrap/dist/css/bootstrap.min.css";
 import {Link} from "react-router-dom";
-import React, {useState} from "react";
-import "./patient.css"
-import FormInput from "./FormInput";
-
 
 export const Patient = () =>{
-    const [values,setValues] = useState({
-        FullName:"",
-        DateofBirth:"",
-        Email:"",
-        Password:""
-    })
-
-    const inputs = [
-        {
-            id:1,
-            name:"FullName",
-            type:"text",
-            placeholder:"Enter your fullName",
-            errorMessage:"Username should be 3-16 characters and shouldn't include any special characters!",
-            label:"Full Name :",
-            pattern:"^[A-Za-z0-9]{3,16}$",
-            required:true,
-         },
-         {
-             id:2,
-             name:"DateofBirth",
-             type:"date",
-             placeholder:"Enter your DOB",
-             errorMessage:"",
-             label:"Date of Birth :"
-
-         },
-         {
-             id:3,
-             name:"Email",
-             type:"Email",
-             placeholder:"Enter your Email",
-             errorMessage:"It should be a valid email address!",
-             label:"Email :",
-             required:true,
-         },
-         {
-             id:4,
-             name:"Password",
-             type:"password",
-             placeholder:"Enter your password",
-             errorMessage:"Password should be 8-10 characters and include at least 1 letter,1 number and 1 special character! ",
-             label:"Password :",
-             pattern: `^(?=.*[0-9])(?=.*[a-zA-Z])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]{8,10}$`,
-             required:true,
-         },
-    ];
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-    };
-
-    const onChange = (e) => {
-        setValues({...values, [e.target.name]: e.target.value});
-    };
-
-
-
-        console.log(values);
-    let Linking;
-    return (
-        <div className="patient">
-        <form onSubmit={handleSubmit}>
-            <h1 className="P1">Get Started!</h1>
-            <h2 className="P2">Patient</h2>
-            <h6 className="P3">Already have an account ?</h6>
-            <Link className="Link" to='/Home'>Sign in</Link>
-            {inputs.map((input) => (
-                <FormInput
-                    key={input.id}
-                    {...input}
-                    value= {values[input.name]}
-                    onChange={onChange}
+    return(
+        <div className="container-fluid">
+      <div className="row">
+        <div className="col-md-6 p-0">
+          <img src={loginImg} alt="example image" className="w-100 h-100" />
+        </div>
+        <div className="col-md-6 d-flex  justify-content-center form-container">
+          
+          <form className="login-form">
+          <div className="login-name">
+                <h2 className="mb-4">Patient Login</h2>
+                </div>
+            <div className="form-group mb-3 row">
+              <label htmlFor="doctorId" className="col-sm-1 col-form-label">
+                Username:
+              </label>
+              <div className="col-sm-10 mt-2 inputLabel">
+                <input
+                  type="text"
+                  id="doctorId"
+                  name="doctorId"
+                  className="form-control"
+                  required
                 />
-                ))}
-            <button className="BTN">SIGN UP</button>
-
-
-        </form>
-    </div>);
-
-
+              </div>
+            </div>
+            <div className="form-group mb-3 row">
+              <label htmlFor="password" className="col-sm-1 col-form-label">
+                Password:
+              </label>
+              <div className="col-sm-10 mt-2 inputLabel">
+                <input
+                  type="password"
+                  id="password"
+                  name="password"
+                  className="form-control"
+                  required
+                />
+              </div>
+            </div>
+            <div className="loging-btn-text">
+            <button type="submit" className="dloging-btn">
+              LOGIN
+            </button>
+            <div className="forgot-password">
+            <p>Forgot Password</p>
+            </div>
+            <div className="register-here">
+            <p>Don't have an account? <Link to="/patient-reg" className="register-hereLink">REGISTER HERE</Link></p>
+            </div>
+            </div>
+          </form>
+          
+        </div>
+      </div>
+    </div>
+  );
 }
